+++ Userland/SampleCodeModule/include/lib.h
@@ -37,7 +37,7 @@ int getC(char * c);
 void putC(char c, uint32_t hexColor);
 
 // Prints the integer number passed as the first argument in the color specified by the second argument.
-void put_int(int num, uint32_t hexColor);
+void put_int(uint64_t num, uint32_t hexColor);
 
 // Prints the formatted string passed as the second argument in the color specified by the hexadecimal color code passed as the first argument. 
 // Similar to c standarlibrary functino printf.
